Respond to Leave to Appeal Journey Map
